Refactor and add persistance of search parameters
